fix: SystemDebug compile errors when DEBUG_MONITOR_ENABLE=1
Fixes:
- rtc_sleep.h: add extern declaration for g_stLowPowerRtcStatus
- rtc_sleep.c: remove static from g_stLowPowerRtcStatus
- SystemDebug.h: afe_bstatus1/afe_bstatus3/afe_fault1 → uint8_t
  (MTP_REG_BSTATUSx.all is UINT8, not UINT16)
- SystemDebug.c: use .all for REG_BSTATUS1/REG_BSTATUS3 access
- SystemDebug.c: afe_cur_raw from SH367309_Read_AFE1.u16Current
- SystemDebug.c: afe_fault1 from REG_BSTATUS1.all (no REG_BFLAG1 field)
- ADC.h: un-comment g_u16ADCValFilter extern
- Keil project: add SystemDebug.c to Source group

Build: FD_Release with DEBUG_MONITOR_ENABLE=1 → 0 Error, 3 Warning

+++ 103 + 309/Project/Source/SystemDebug.h
@@ -79,9 +79,9 @@ struct SYSTEM_DEBUG {
 	uint16_t soc_vtotal;
 
 	/* --- AFE --- */
-	uint16_t afe_bstatus1;
-	uint16_t afe_bstatus3;
-	uint16_t afe_fault1;
+	uint8_t  afe_bstatus1;
+	uint8_t  afe_bstatus3;
+	uint8_t  afe_fault1;
 	uint16_t afe_cur_raw;
 	uint16_t afe_pec_err;
 	uint16_t afe_cell_min_mv;
